now with MOTD dashboard

+++ dashboard.py
@@ -86,18 +86,18 @@ def get_num_plays():
                         num_plays = count_plays.read()
                 return num_plays
 
-def update(video_title):
+def update_dashboard():
 	with open('/tmp/motd.tmp', 'w') as dashboard_text:
 		dashboard_text.truncate()
 		dashboard_text.write('           _             _     \n')
 		dashboard_text.write('  _____  _| |_ ___ _   _| |__  \n')
-		dashboard_text.write(' / _ \\ \\/ / __/ __| | | | '_ \\ \n')
+		dashboard_text.write(' / _ \\ \\/ / __/ __| | | | \'_ \\ \n')
 		dashboard_text.write('|  __/>  <| |_\\__ \\ |_| | |_) |\n')
 		dashboard_text.write(' \\___/_/\\_\\\\__|___/\\__,_|_.__/ \n')
-		dashboard_text.write('Playing: ' + video_title + '\n\n')
-    dashboard_text.write('Launched: ' + get_process_time("start") + '\n')
+		dashboard_text.write('Playing: ' + video_title + '\n')
+    		dashboard_text.write('Launched: ' + get_process_time("start") + '\n')
 		dashboard_text.write('Playing for: ' + get_process_time("duration") + '\n')
-		#dashboard_text.write('Playlist loops: ' + get_num_plays() + '\n')
+		dashboard_text.write('Looped: ' + get_num_plays() + ' times\n\n')
 		
-
-os.system('sudo mv /tmp/motd.tmp /etc/motd')
+	if os.path.isfile('/tmp/motd.tmp'):
+		os.system('sudo mv /tmp/motd.tmp /etc/motd')
